Add SHA256 utility function
This adds a new function in the utils library to calculate the SHA256 hash for a file. 

I propose this addition as there are libraries of hashes that do not support SHA512 hashes, while SHA256 hashes are supported. Generally, my experience has been that SHA256 hashes are more widely used for malware identification, etc.

Inclusion of the function here would allow individual output modules the option of using this hash algorithm, without requiring any changes to existing code.